v.11 - Got projectiles to spawn from muzzle
I was able to set up projectiles to begin spawning from the muzzle in single player.  My sticking point had to do with assigning a static mesh to the ball.  I was assigning a static mesh, but i wasn't attaching it to the root component.  It resulted in the static mesh spawning at the origin and appearing to not move.