v.16 Projectile size and Generate Hit Events
Removed C++ defintions for projectile size and decided to leave them in the Unreal UI so they are easier to edit on the fly.  In addition to this, Enabled "Generate Hit Events" for both the Character capsule component and the projectile so that they can fire the OnHit logic.  Finally, I modified the OnHit logic so it will recognize when the projectile is hitting another player character.