v.20 - Inserting Snapshot Functionality
Setting up the logic for the snapshot spawns.  I incorporated the features necessary to use the TestSnapshot command in the Unreal UI command line to get an exact snapshot of the player to spawn in the world.  Not sure if it works for multiplayer yet.  If I use the StartGame command it will crash because the player list is not populating.

I need to refactor the PlayerList code because I need it to be of data type APC_Player instead of APlayerController.  Therefore, I am making this check in just in case everything goes to hell.