v.20-branch- Incorporated snapshot initial state and post-possession state
Modified gravity and collision profile on the initial character spawn-in, and changed the gravity and collision when a player controller possesses the character.

Since gravity and collision profile aren't replicated properties and only see changes on the server, I used a netmulticast RPC to relay the information down to the clients.